requirements.txt
mistralai
python-dotenv
termcolor